+++ shared/src/commonMain/kotlin/ui/viewmodels/ErrorViewModel.kt
@@ -7,7 +7,6 @@ import at.asitplus.valera.resources.Res
 import at.asitplus.valera.resources.info_text_error_action_reset_app
 import at.asitplus.valera.resources.info_text_error_action_start_screen
 import at.asitplus.valera.resources.info_text_error_cause_reset_app
-import at.asitplus.wallet.app.common.enrichMessage
 import kotlinx.coroutines.runBlocking
 import org.jetbrains.compose.resources.StringResource
 import org.jetbrains.compose.resources.getString
@@ -25,7 +24,7 @@ class ErrorViewModel(
     var textCause: String?
 
     private val displayThrowable = (throwable as? ErrorHandlingOverrideException)?.cause ?: throwable
-    val message = displayThrowable.enrichMessage()
+    val message = displayThrowable.message ?: displayThrowable.toString()
     val cause = displayThrowable.cause?.toString()
     private val isAppResetRequired = message == AppResetRequiredException.toString()